+++ CHANGELOG.md
@@ -7,7 +7,8 @@
 
 ### Bug Fixes
 
--   Fix crash when converting `export default undefined`, #2175
+-   Fix crash when converting `export default undefined`, #2175.
+-   Fix error in console when clicking on headings in the readme, #2170.
 
 ### Thanks!